shoppingApp/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
PyMySQL==1.0.2
django-cors-headers

